+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_validators.py
@@ -713,6 +713,7 @@ def process_nic_create_namespace(cmd, namespace):
 def process_public_ip_create_namespace(cmd, namespace):
     get_default_location_from_resource_group(cmd, namespace)
     validate_public_ip_prefix(cmd, namespace)
+    validate_ip_tags(cmd, namespace)
     validate_tags(namespace)