+++ src/command_modules/azure-cli-monitor/azure/cli/command_modules/monitor/tests/latest/test_monitor_unittest.py
@@ -88,3 +88,66 @@ class MonitorNameOrIdTest(unittest.TestCase):
         self.assertFalse(hasattr(ns, 'namespace'))
         self.assertFalse(hasattr(ns, 'parent'))
         self.assertFalse(hasattr(ns, 'resource_type'))
+
+
+class MonitorMetricAlertActionTest(unittest.TestCase):
+
+    def _build_namespace(self, name_or_id=None, resource_group=None, provider_namespace=None, parent=None,
+                         resource_type=None):
+        from argparse import Namespace
+        return Namespace()
+
+    def call_condition(self, ns, value):
+        from azure.cli.command_modules.monitor.actions import MetricAlertConditionAction
+        MetricAlertConditionAction('--condition', 'condition').__call__(None, ns, value.split(), '--condition')
+
+    def check_condition(self, ns, time_aggregation, metric_namespace, metric_name, operator, threshold):
+        prop = ns.condition[0]
+        self.assertEqual(prop.time_aggregation, time_aggregation)
+        self.assertEqual(prop.metric_name, metric_name)
+        self.assertEqual(prop.operator, operator)
+        self.assertEqual(prop.threshold, threshold)
+        self.assertEqual(prop.metric_namespace, metric_namespace)
+
+    def check_dimension(self, ns, index, name, operator, values):
+        dim = ns.condition[0].dimensions[index]
+        self.assertEqual(dim.name, name)
+        self.assertEqual(dim.operator, operator)
+        self.assertEqual(dim.values, values)
+
+    def test_monitor_metric_alert_condition_action(self):
+
+        from knack.util import CLIError
+
+        ns = self._build_namespace()
+        self.call_condition(ns, 'avg ns."CPU Percent" > 90')
+        self.check_condition(ns, 'Average', 'ns', 'CPU Percent', 'GreaterThan', '90')
+
+        ns = self._build_namespace()
+        self.call_condition(ns, 'avg ns."a.b/c_d" > 90')
+        self.check_condition(ns, 'Average', 'ns', 'a.b/c_d', 'GreaterThan', '90')
+
+        ns = self._build_namespace()
+        self.call_condition(ns, 'avg CPU Percent > 90')
+        self.check_condition(ns, 'Average', None, 'CPU Percent', 'GreaterThan', '90')
+
+        ns = self._build_namespace()
+        self.call_condition(ns, 'avg "a.b/c_d" > 90')
+        self.check_condition(ns, 'Average', None, 'a.b/c_d', 'GreaterThan', '90')
+
+        ns = self._build_namespace()
+        self.call_condition(ns, 'avg SuccessE2ELatency > 250 where ApiName includes GetBlob or PutBlob')
+        self.check_condition(ns, 'Average', None, 'SuccessE2ELatency', 'GreaterThan', '250')
+        self.check_dimension(ns, 0, 'ApiName', 'includes', ['GetBlob', 'PutBlob'])
+
+        ns = self._build_namespace()
+        self.call_condition(ns, 'avg ns.foo/bar_doo > 90')
+        self.check_condition(ns, 'Average', 'ns', 'foo/bar_doo', 'GreaterThan', '90')
+
+        ns = self._build_namespace()
+        with self.assertRaisesRegexp(CLIError, 'usage error: --condition'):
+            self.call_condition(ns, 'avg blah"what > 90')
+
+        ns = self._build_namespace()
+        with self.assertRaisesRegexp(CLIError, 'usage error: --condition'):
+            self.call_condition(ns, 'avg Wra!!ga * woo')