+++ src/command_modules/azure-cli-resource/azure/cli/command_modules/resource/commands.py
@@ -241,15 +241,15 @@ def load_command_table(self, _):
 
     with self.command_group('policy definition', resource_policy_definitions_sdk, resource_type=ResourceType.MGMT_RESOURCE_POLICY) as g:
         g.custom_command('create', 'create_policy_definition')
-        g.command('delete', 'delete')
-        g.command('list', 'list')
+        g.custom_command('delete', 'delete_policy_definition')
+        g.custom_command('list', 'list_policy_definition')
         g.custom_show_command('show', 'get_policy_definition')
-        g.generic_update_command('update', custom_func_name='update_policy_definition', custom_func_type=resource_custom)
+        g.custom_command('update', 'update_policy_definition')
 
     with self.command_group('policy set-definition', resource_policy_set_definitions_sdk, resource_type=ResourceType.MGMT_RESOURCE_POLICY, min_api='2017-06-01-preview') as g:
         g.custom_command('create', 'create_policy_setdefinition')
-        g.command('delete', 'delete')
-        g.command('list', 'list')
+        g.custom_command('delete', 'delete_policy_setdefinition')
+        g.custom_command('list', 'list_policy_setdefinition')
         g.custom_show_command('show', 'get_policy_setdefinition')
         g.custom_command('update', 'update_policy_setdefinition')