+++ src/command_modules/azure-cli-network/HISTORY.rst
@@ -3,6 +3,12 @@
 Release History
 ===============
 
+2.2.8
++++++
+* Deprecated `network interface-endpoint` command names in favor of `network private-endpoint`.
+* `express-route peering connection create`: Fix issue where `--peer-circuit` would not accept an ID.
+* `public-ip create`: Fix issue where `--ip-tags` did not work correctly.
+
 2.2.7
 +++++
 * `nic create` - Add `--app-gateway-address-pools` and `--gateway-name` arguments to support adding application